+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The following emojis are used to highlight certain changes:
 ### Added
 
 - Added endpoints to show and purge connected peers [#194](https://github.com/ipfs/rainbow/pull/194)
+- Added flags to configure bitswap tuning params:
+  - `bitswap-max-requests`
+  - `bitswap-max-providers`
+  - `bitswap-max-timeout`
 
 ### Changed